pyproject.toml
[tool.poetry]
name = "annotation-app"
version = "0.1.0"
description = ""
authors = ["Eiji Takahashi <comu2e@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
streamlit = "^1.32.1"
dataset = "^1.6.2"
ja-sentence-segmenter = "^0.0.2"
streamlit-shortcuts = "^0.1.1"
streamlit-authenticator = "^0.3.1"
boto3 = "^1.34.64"
stqdm = "^0.0.5"
mecab = "^0.996.3"
hojichar = "^0.9.0"
tqdm = "^4.66.2"
pymysql = "^1.1.0"
google = "^3.0.0"
cloud-sql-python-connector = "^1.8.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
